Submitted by:	Randy Terbush and Howard Fear Fergot to put in the -Wall patch - chuck
git-svn-id: https://svn.apache.org/repos/asf/httpd/httpd/branches/1.3.x@76547 13f79535-47bb-0310-9956-ffa450edef68

+++ src/modules/standard/mod_include.c
@@ -99,7 +99,7 @@ void add_include_vars(request_rec *r, char *timefmt)
       table_set(e, "USER_NAME", pw->pw_name);
     } else {
       char uid[16];
-      sprintf(uid, "user#%ld", r->finfo.st_uid);
+      sprintf(uid, "user#%ld", (unsigned long)r->finfo.st_uid);
       table_set(e, "USER_NAME", uid);
     }
